requirements.txt
# Backend Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.23
aiosqlite>=0.19.0

# Document Processing
PyPDF2>=3.0.1
pdfminer.six>=20221105
python-docx>=1.1.0

# NLP
spacy>=3.7.2
nltk>=3.8.1
sentence-transformers>=2.2.2

# Machine Learning
scikit-learn>=1.3.2
numpy>=1.26.2
pandas>=2.1.3
joblib>=1.3.2

# Data Visualization (for notebooks)
matplotlib>=3.8.2
seaborn>=0.13.0

# Development & Testing
jupyter>=1.0.0
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-asyncio>=0.21.1
httpx>=0.25.2

# Code Quality
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0

# Logging
loguru>=0.7.2
